requirements.txt
# Flask web framework
Flask==2.3.3

# Instagram downloading
instaloader==4.10.3

# YouTube API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Gemini AI
google-generativeai==0.3.2

# Video processing
opencv-python==4.8.1.78
Pillow==10.0.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Additional required packages
urllib3==1.26.18
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
gunicorn
